draft unicoderange implementation
